namah service file strutcture update
