+++ services/namah/services/translateObjectListKeys.js
@@ -1,0 +1,14 @@
+//Services
+const translateObjectKeys = require('.').translateObjectKeys;
+
+const translateObjectListKeys = (targetObjectList) => {
+    const newObjectList = new Array(targetObjectList.length);
+
+    targetObjectList.forEach((currentObject, index) => {
+        newObjectList[index] = translateObjectKeys(currentObject);
+    });
+
+    return newObjectList;
+};
+
+module.exports = translateObjectListKeys;
